update routes improvement

+++ api/routes/backpacking_trip_routes/backpacking_trip_common_cities.js
@@ -83,25 +83,46 @@ router.post('/post_backpacking_trip_common_city',upload.single('common_city_imag
     });
 });
 
-router.patch('/update_backpacking_trip_common_city',upload.single('common_city_image'),(req,res)=>{
+//update
+router.post('/update_backpacking_trip_common_city',upload.single('common_city_image'),(req,res)=>{
 
     Backpacking_Trip_Common_City.findOne({_id:req.body.backpacking_trip_common_city_id})
     .exec()
     .then(result=>{
-        const filePath=String(path.dirname(require.main.filename))+'\\'+String(result.common_city_image_path);
-        console.log(filePath);
-        fs.unlinkSync(`${filePath}`);
-        console.log('old state image deleted');
-        Backpacking_Trip_Common_City.updateOne({_id:req.body.backpacking_trip_common_city_id},{
-            common_city_image_path:req.file.path,
-            is_available:req.body.is_available
-        })
-        .exec()
-        .then(result1=>{
-            res.json({
-                message:'Data updated'
+        if(req.file==undefined)
+        {
+            Backpacking_Trip_Common_City.updateOne({_id:req.body.backpacking_trip_common_city_id},{
+                common_city_image_path:req.file.path,
+                is_available:req.body.is_available
+            })
+            .exec()
+            .then(result1=>{
+                res.json({
+                    message:'Data updated'
+                });
             });
-        });
+
+        }
+
+        else{
+            const filePath=String(path.dirname(require.main.filename))+'\\'+String(result.common_city_image_path);
+            console.log(filePath);
+            fs.unlinkSync(`${filePath}`);
+            console.log('old state image deleted');
+            Backpacking_Trip_Common_City.updateOne({_id:req.body.backpacking_trip_common_city_id},{
+                common_city_image_path:req.file.path,
+                is_available:req.body.is_available
+            })
+            .exec()
+            .then(result1=>{
+                res.json({
+                    message:'Data updated'
+                });
+            });
+
+
+        }
+        
     })
     .catch(err=>{
         res.json({
